refactor: give cargo-budget-report's numbered modules meaningful names
Rename the seven module_N files in cargo-budget-report/src/ after what
they contain, matching the naming style of the already-named modules:

- module_2.rs  -> limit_checks.rs          (budget limit bound checks)
- module_3.rs  -> config.rs                (Config + TOML parse_config)
- module_4.rs  -> url_checks.rs            (URL scheme/host predicates)
- module_10.rs -> error.rs                 (Error, Result, Simulation*)
- module_8.rs  -> edge_case_tests.rs       (off-by-one/zero-length tests)
- module_18.rs -> boundary_tests.rs        (boundary-condition tests)
- module_28.rs -> additional_edge_tests.rs (additional edge-case tests)

Import/mod-declaration fixups in main.rs and derive.rs, plus doc-comment
cross-references. No behaviour changes.

Note: additional_edge_tests.rs was already orphaned before this change
(never declared via 'mod'), so it stays unwired; renaming only.
