+++ cargo-budget-report/src/edge_case_tests.rs
